Fix compilation: conflicting types for ‘netlog_gperf_lookup’
```
gperf < src/netlog-gperf.gperf > src/netlog-gperf.c
  CC       src/netlog-gperf.o
src/netlog-gperf.c:34:1: error: conflicting types for ‘netlog_gperf_lookup’
 netlog_gperf_lookup (register const char *str, register size_t len)
 ^~~~~~~~~~~~~~~~~~~
In file included from src/netlog-gperf.c:8:0:
src/netlog-conf.h:28:30: note: previous declaration of ‘netlog_gperf_lookup’ was here
 const struct ConfigPerfItem* netlog_gperf_lookup(const char *key, unsigned length);
                              ^~~~~~~~~~~~~~~~~~~
make[2]: *** [Makefile:932: src/netlog-gperf.o] Error 1
```

+++ src/netlog-conf.h
@@ -25,7 +25,7 @@
 #include "conf-parser.h"
 #include "netlog-manager.h"
 
-const struct ConfigPerfItem* netlog_gperf_lookup(const char *key, unsigned length);
+const struct ConfigPerfItem* netlog_gperf_lookup(const char *key, size_t length);
 int config_parse_netlog_remote_address(const char *unit,
                                        const char *filename,
                                        unsigned line,
